Remember scroll position when navigating back after reading an article from Brave Today on NTP
Stores article url in window.history.state. When the back is loaded and has that state, we scroll down to an element that is rendered and is displaying the article url which was previously visited.
TODO: do this for list-type cards. Problem with those is that we must wait for all previous images to be loaded so that the correct scroll position is not maintained as item heights increase when images are loaded.